requirements.txt
azure-common==1.1.3
azure-nspkg==1.0.0
azure-storage-blob==12.5.0
flask>=0.12.3
Flask-Cors==3.0.10
pylint==2.6.0
waitress==1.0.1
futures==3.0.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.3
requests>=2.20.0
six==1.10.0
Werkzeug==0.11.5
mysql-connector-python==8.0.23
python-dotenv==0.15.0